Add test for throwing error when trying to finalize purchase with empty cart
The code changes include the addition of a new test case that checks whether an error is thrown when trying to finalize a purchase with an empty cart. The test case creates a new instance of the "Carrinho" class, calls the "finalizaCompra" method on it, and expects an error with the message "Carrinho de compras vazio" to be thrown.